Move similarities and Remove SegmentResponse files
The similarity identification is bulky inside of an already dense
heat_transfer module. Similarity has to do with positions of the
boreholes, so it has been moved to the boreholes module.

The SegmentResponse module was initially created when I was considering
hash tables. However, hash tables are increadibly slow when compared to
indexing std::vector's. The SegmentResponse struct should work fine
inside of heat_transfer.