Fixes bug pointed out by Matt Mitchell
this is the actual fix for #35, the previous fix was only temporary and
did not resolve the actual bug at hand.

related to #38

Matt makes use of a constant log time expansion of 0.5. The same
issue Matt was seeing was showing up in the g-function adaptive test
example. The tests ran for comparisons have made use of Eskilson's
original 27 time steps. The time difference (delta_t) between the first
two points were greater than the time in seconds of ln(t/ts)=-8.5.
There was an if statement in interp1d that had not been hit by any of
the prior tests. This has been patched.

It also appears that when this case occurred that the load history was
not properly reconstructed. The load history needed to have a
reconstruction of p+1, rather than just p.

Thanks, Matt.