Make library rely on BUILD_SHARED_LIBS
"
do we have a need for this to be explicitly SHARED? If the goal is to
always link it to energypluslib, then STATIC would make more sense.
But I would omit the keyword, rely on the value of BUILD_SHARED_LIBS to
determine if STATIC or SHARED is picked.
" -@jmarrec